Modify the parsing method in spider
Switch extract() with get() referring to user and date.

Since extract() can return a list of strings and the user and the date needs  only a string, it's better using the get() which return only one string.

Parse the date variable and format it in a better way using the datetime module.