Add a db handler class
Create a singleton class to universally manage the db in the project.

So, update the pipelines to use the db manager methods to storage the reviews.

Rename the fields of the item container to match with those in the spider, since filmID and filmTitle are used instead of film_id and film_title.