Update the db and the pipeline modules
Modify the DBManager methods and, therfore, process_item() in pipeline.

Document and specify the type of arguments and return value of the methods of DBManager.

As getReviewsOf() tried to get access to (and return) the reviews through the attribute 'reviews' of the film selected by 'filmID', this always raised an error, because the reference is non-existent outside the session. Get rid of that by selecting directly the reviews (instead of the film) filtered by 'filmID' and returning them.

Rename getFilmBy() with getFilmByID(), so that it's more self-explanatory.

Add the class attribute film in the pipeline to enhance the performances, so that it's not needed to query each time the film from the db, which the review to be added  must refer to.